requirements.txt
arxiv==2.1.3
numpy==2.4.2
openai==1.99.9
pypdf==5.3.0
python-dotenv==1.0.1
streamlit==1.54.0
tqdm==4.67.1
